Storing the collision index inside of the GameObject itself was dumb :)
Storing the drawIndex inside of the GameObject itself was dumb :)

The position at which each vertex is initialized doesn't actually matter. On the next frame, they will be moved wherever they need to be.

A proper usage of the custom vTables for GameObjects, with different IDs (indexes in the vTable) for different functions.

Putting flags inside of the GameObject itself was dumb :)
With these and other changes, I feel I am approaching an Entity-Component System sort of deal.

Now you can (sort of) easily write in your own update methods, all in accordance with the custom vTables.

fixed a line

Removed unused variable.